# sql: implement the PostgreSQL wire protocol