--- a/sql/distinct.go
+++ b/sql/distinct.go
@@ -146,7 +146,7 @@ func (n *distinctNode) ExplainPlan() (string, string, []planNode) {
 		strs := make([]string, 0, len(columns))
 		for i, column := range columns {
 			if n.columnsInOrder[i] {
-				strs = append(strs, column)
+				strs = append(strs, column.name)
 			}
 		}
 		description = strings.Join(strs, ",")