--- a/sql/pgwire/context.go
+++ b/sql/pgwire/context.go
@@ -1,0 +1,31 @@
+// Copyright 2015 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License. See the AUTHORS file
+// for names of contributors.
+//
+// Author: Ben Darnell
+
+package pgwire
+
+import (
+	"github.com/cockroachdb/cockroach/base"
+	"github.com/cockroachdb/cockroach/sql"
+	"github.com/cockroachdb/cockroach/util/stop"
+)
+
+// Context holds parameters needed to setup a postgres-compatible server.
+type Context struct {
+	*base.Context
+	Executor *sql.Executor
+	Stopper  *stop.Stopper
+}