--- a/sql/plan.go
+++ b/sql/plan.go
@@ -253,11 +253,17 @@ func (p *planner) releaseLeases(db client.DB) {
 	}
 }
 
+type column struct {
+	name string
+	typ  parser.Datum
+}
+
 // planNode defines the interface for executing a query or portion of a query.
 type planNode interface {
-	// Columns returns the column names. The length of the returned slice is
-	// guaranteed to be equal to the length of the tuple returned by Values().
-	Columns() []string
+	// Columns returns the column names and types . The length of the
+	// returned slice is guaranteed to be equal to the length of the
+	// tuple returned by Values().
+	Columns() []column
 	// The indexes of the columns the output is ordered by. Indexes are 1-based
 	// and negative indexes indicate descending ordering. The ordering return
 	// value may be nil if no ordering has been performed. The prefix return