Out for early review per @petermattis.

Transactions are not yet implemented, and authentication is incomplete.

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/3218)
<!-- Reviewable:end -->