@tamird 2015-11-24
OK, this should be passing now. Note that the last commit represents the last outstanding bit of work (implement type checking for subqueries, or equivalent).

Nevertheless, this is ready for review.