@tamird 2015-11-24
OK, subqueries should work now as well.