@petermattis 2015-11-24
Review status: 0 of 29 files reviewed at latest revision, 9 unresolved discussions, some commit checks failed.

---

*[rpc/codec/conn.go, line 36 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vWGN5UOK-8sYnOSAx:-K3vWGN5UOK-8sYnOSAy:-1181639181) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/rpc/codec/conn.go#L36)):*
Let's revert this for now.

---

*[server/server.go, line 221 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vWvVpV6i0ENQYPuKR:-K3vWvVpV6i0ENQYPuKS:235780252) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/server/server.go#L221)):*
Perhaps add a `pgaddr` flag. 

---

*[sql/driver/wire.proto, line 99 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vXVJgGQZsKVgCwfLv:-K3vXVJgGQZsKVgCwfLw:1698636128) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/driver/wire.proto#L99)):*
Any reason for changing the field id to `3`?

---

*[sql/logic_test.go, line 466 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vYCWoYuYw_D9mwJTu:-K3vYCWoYuYw_D9mwJTv:1453120591) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/logic_test.go#L466)):*
What does `%q` do if the argument is not a string/[]byte?

---

*[sql/logic_test.go, line 480 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vYQZGM2LqXzxeeCzh:-K3vYQZGM2LqXzxeeCzi:-1497524547) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/logic_test.go#L480)):*
Can you add a comment explaining what this is for? You're converting the `[]byte` to a string, but that's different than `fmt.Sprint()`?

---

*[sql/parser/values.go, line 36 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vYhDf3SD53zIUnTbF:-K3vYhDf3SD53zIUnTbG:-143753303) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/parser/values.go#L36)):*
Any reason for the changes here?

---

*[sql/scan.go, line 530 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vZWDyE0h83-EUaaEC:-K3vZWDyE0h83-EUaaED:-1688613307) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/scan.go#L530)):*
Type checking twice? Probably deserves a comment to explain that we have to type check the normalized expression to determine its output type.

---

*[sql/testdata/datetime, line 20 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vZvGtKwBKVQmn_9Fg:-K3vZvGtKwBKVQmn_9Fh:790011992) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/testdata/datetime#L20)):*
Why did this change? Pgwire might be sending the dates differently, but they still arrive as `time.Time`, right?

---

*[sql/testdata/txn, line 77 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3v_862btq4MMTCq30g:-K3v_862btq4MMTCq30h:2145507833) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/testdata/txn#L77)):*
Why is this being output? Are we outputting all of the results over pgwire or only the last one?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218)*
<!-- Sent from Reviewable.io -->