@petermattis 2015-11-24
Can you include the output of the new `Select1` variations?

---

Review status: 0 of 29 files reviewed at latest revision, 9 unresolved discussions, some commit checks failed.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-:-K3v_gQuwEH44NsWGSQd:1401747124)*
<!-- Sent from Reviewable.io -->