@tamird 2015-11-24
Review status: 0 of 28 files reviewed at latest revision, 9 unresolved discussions.

---

*[sql/driver/wire.proto, line 99 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vXVJgGQZsKVgCwfLv:-K3vanEZAQmwfNbNrZOq:529426274) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/driver/wire.proto#L99)):*
Done goofed. Reverted.

---

*[sql/logic_test.go, line 466 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vYCWoYuYw_D9mwJTu:-K3vas-CE7qo6toffzan:53268199) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/logic_test.go#L466)):*
actually, I'm not sure.

---

*[sql/logic_test.go, line 480 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vYQZGM2LqXzxeeCzh:-K3vblk_iKZgP3p9F5Pj:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/logic_test.go#L480)):*
Done.

---

*[sql/parser/values.go, line 36 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vYhDf3SD53zIUnTbF:-K3vbmj4TinxfIDwz-pi:700048134) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/parser/values.go#L36)):*
Nah, just seemed simpler. It's in a separate commit.

---

*[sql/scan.go, line 530 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vZWDyE0h83-EUaaEC:-K3vce5sBYtiQaoNR2lR:1256157153) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/scan.go#L530)):*
Turns out it's possible to type check just once. Changed.

---

*[sql/testdata/datetime, line 20 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vZvGtKwBKVQmn_9Fg:-K3vch5x2jccZlGtKc0H:428150391) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/testdata/datetime#L20)):*
They were previously arriving as `driver.Date`, not `time.Time`

---

*[sql/testdata/txn, line 77 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3v_862btq4MMTCq30g:-K3vcoqzSaU5dxYKwMeS:1806070135) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/sql/testdata/txn#L77)):*
That's the behaviour change (see one of the commit messages). pgwire is outputting them all.

---

*[rpc/codec/conn.go, line 36 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-K3vWGN5UOK-8sYnOSAx:-K3va_6zAhhgCi0B60xr:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/66ca88ba4b8dc001ed841ce1da8230854453ab60/rpc/codec/conn.go#L36)):*
Done.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218)*
<!-- Sent from Reviewable.io -->