@tamird 2015-11-24
Sure. For now, here they are:
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkBank-4              1046735       787929        -24.73%
BenchmarkSelect1_HTTP-4      490991        484219        -1.38%
BenchmarkSelect1_HTTPS-4     470506        519195        +10.35%
BenchmarkSelect1_RPC-4       121440        130514        +7.47%
BenchmarkSelect1_RPCS-4      142458        142702        +0.17%
```

---

Review status: 0 of 28 files reviewed at latest revision, 9 unresolved discussions, some commit checks failed.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/3218#-:-K3vpZFC_p4G4mi3Cd2H:-1500686396)*
<!-- Sent from Reviewable.io -->